Add Setup API config screen and modify API config screen, now every request uses saved api ip-address in shared preferences

+++ mobile_app/lib/screens/manage_device_screen.dart
@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:http/http.dart' as http;
 import '../models/device.dart';
 import 'streaming_screen.dart';
 import 'edit_device_screen.dart';
+import 'main_screen.dart';
 
 import '../api_constants.dart';
+
 
 class ManageDeviceScreen extends StatelessWidget {
   final Device device;
@@ -12,9 +15,20 @@ class ManageDeviceScreen extends StatelessWidget {
   const ManageDeviceScreen({super.key, required this.device});
 
   Future<void> _deleteDevice(BuildContext context) async {
-    final url = Uri.http(baseUrl, "${deleteDeviceEndpoint}${device.id}");
 
     try {
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      String? baseUrl = prefs.getString('api-ip');
+
+      if (baseUrl == null || baseUrl.isEmpty) {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => MainScreen()),
+        );
+        return;
+      }
+
+      final url = Uri.http(baseUrl, "${deleteDeviceEndpoint}${device.id}");
       final response = await http.delete(url);
 
       // zmienic
@@ -24,7 +38,6 @@ class ManageDeviceScreen extends StatelessWidget {
         );
         Navigator.pop(context);
       } else {
-        print(url);
         ScaffoldMessenger.of(context).showSnackBar(
           SnackBar(content: Text('Failed to delete the device.')),
         );
